Trait relay_utils::metrics::StandaloneMetrics

pub trait StandaloneMetrics: Metrics {
    fn update<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn update_interval(&self) -> Duration;

    fn spawn(self) { ... }
}

Standalone metrics API.

Metrics of this kind know how to update themselves, so we may just spawn and forget the asynchronous self-update task.

Required Methods

fn update<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Update metric values.

fn update_interval(&self) -> Duration

Metrics update interval.

Provided Methods

fn spawn(self)

Spawn the self update task that will keep update metric value at given intervals.