Module relay_utils::relay_loop

Structs

Loop

Generic relay loop.

LoopMetrics

Relay loop metrics builder.

Constants

RECONNECT_DELAY

Default pause between reconnect attempts.

Traits

Client

Basic blockchain client from relay perspective.

Functions

reconnect_failed_client

Deal with the client who has returned connection error.

relay_loop

Returns generic loop that may be customized and started.

relay_metrics

Returns generic relay loop metrics that may be customized and used in one or several relay loops.