Module relay_utils::metrics

Structs

Counter

The underlying implementation for [Counter] and [IntCounter].

F64

A atomic float.

FloatJsonValueMetric

Metric that represents float value received from HTTP service as float gauge.

Gauge

The underlying implementation for [Gauge] and [IntGauge].

GlobalMetrics

Global Prometheus metrics.

I64

A atomic signed integer.

MetricsAddress

Unparsed address that needs to be used to expose Prometheus metrics.

MetricsParams

Prometheus endpoint MetricsParams.

Opts

A struct that bundles the options for creating most [Metric] types.

Registry

A struct for registering Prometheus collectors, collecting their metrics, and gathering them into MetricFamilies for exposition.

U64

A atomic unsigned integer.

Enums

PrometheusError

The error types for prometheus.

Traits

Atomic

An interface for atomics. Used to generically model float metrics and integer metrics, i.e. Counter and IntCounter.

Collector

An interface for collecting metrics.

Metric

Metric API.

StandaloneMetric

Standalone metric API.

Functions

metric_name

Returns metric name optionally prefixed with given prefix.

register

set_gauge_value

Set value of gauge metric.

Type Aliases

CounterVec

The underlying implementation for [CounterVec] and [IntCounterVec].

F64SharedRef

Shared reference to f64 value that is updated by the metric.

GaugeVec

The underlying implementation for [GaugeVec] and [IntGaugeVec].

IntGauge

Int gauge metric type.